+++ packages/0x.js/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+vx.x.x
+------------------------
+    * Make `DecodedLogEvent<A>` contain `LogWithDecodedArgs<A>` under log key instead of merging it in like web3 does (#234)
+    * Modify order validation methods to validate against the `latest` block, not against the `pending` block (#236)
+
 v0.26.0
 ------------------------
     * Add post-formatter for logs converting `blockNumber`, `logIndex`, `transactionIndex` from hexes to numbers (#231)